Added minimap
