+++ Assets/Script/Photon/GameManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8ae7e0bda8be7b449828b8ad799c0919
+guid: 1c06b15fdab6d8a47ae4c5aa574d3d30
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
