+++ Assets/Script/Track/Minimap.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1c06b15fdab6d8a47ae4c5aa574d3d30
+guid: 628a6f3488d934d4a83988a3b4b72a95
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
