+++ Assets/Script/Track/Minimap.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8a7a852aa3afbad42a9f6ffcb1fea56a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
