+++ Assets/Script/Track/Minimap/LocalPlayer.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bd0a48ea8f1f5844e8688360f9725a86
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
