+++ Assets/Script/Track/Minimap/Minimap.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c39857ad043eb4147886d627490f022a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
