+++ Assets/Script/Track/Minimap/Minimap.renderTexture.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e5ee77acffaef404f98854bf99b5e20e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
